Move virtual env create logic into installer
